chore!: use pure ES modules
