--- ava.config.js
@@ -1,3 +1,0 @@
-export default {
-  verbose: true,
-}
