--- test/fixtures/_valid.js
@@ -1,3 +1,0 @@
-const test = true
-
-module.exports = { test }
