+++ test/fixtures/_valid.mjs
@@ -1,0 +1,1 @@
+export const test = true
